+++ pom.xml
@@ -1,0 +1,96 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>ud.ninia.prog3-proyecto</groupId>
+    <artifactId>prog3-proyecto</artifactId>
+    <version>1.0-SNAPSHOT</version>
+
+
+    <licenses>
+        <license>
+            <name>GNU General Public License, Version 3.0</name>
+            <url>https://github.com/Ninia/prog3-proyecto/blob/master/LICENSE</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+
+    <contributors>
+
+        <contributor>
+            <name>Rafael Romon</name>
+            <email>rafaelromonsagredo@opendeusto.es</email>
+            <url>https://github.com/rafaelromon</url>
+            <organization>Ninia</organization>
+            <organizationUrl>https://github.com/Ninia</organizationUrl>
+            <roles>
+                <role>developer</role>
+            </roles>
+            <timezone>Europe/Madrid</timezone>
+        </contributor>
+
+        <contributor>
+            <name>Saul Segura</name>
+            <email>saul.segura@opendeusto.es </email>
+            <url>https://github.com/luasaul</url>
+            <roles>
+                <role>developer</role>
+            </roles>
+            <timezone>Europe/Madrid</timezone>
+        </contributor>
+
+        <contributor>
+            <name>Mikel Solabarrieta</name>
+            <email>mikelsr@opendeusto.es</email>
+            <url>https://github.com/mikelsr</url>
+            <organization>Ninia</organization>
+            <organizationUrl>https://github.com/Ninia</organizationUrl>
+            <roles>
+                <role>developer</role>
+            </roles>
+            <timezone>Europe/Madrid</timezone>
+        </contributor>
+
+    </contributors>
+
+
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.ftpserver</groupId>
+            <artifactId>ftpserver-core</artifactId>
+            <version>1.1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>20160810</version>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+        </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest-core</artifactId>
+            <version>1.3</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
+
+</project>
