+++ src/main/java/ud/ninia/prog3_proyecto/influxdb/monitor/CPU.java
@@ -1,0 +1,30 @@
+package ud.ninia.prog3_proyecto.influxdb.monitor;
+
+import ud.ninia.prog3_proyecto.utils.posix.ShellCommand;
+
+public class CPU {
+
+    public static double getTemperature() {
+        return Double.parseDouble(
+                ShellCommand.executeCommand("/usr/local/proyectopbd/cpu-temperature.sh")
+        );
+    }
+
+    public static double getUsage() {
+        return Double.parseDouble(
+                ShellCommand.executeCommand("/usr/local/proyectopbd/cpu-usage.sh")
+        );
+    }
+
+    public static void main(String[] args) {
+
+        while (true) {
+            System.out.println("CPU usage: " + CPU.getUsage());
+            try {
+                Thread.sleep(1000);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+    }
+}
