+++ src/main/java/ud/ninia/prog3_proyecto/influxdb/monitor/Memory.java
@@ -1,0 +1,24 @@
+package ud.ninia.prog3_proyecto.influxdb.monitor;
+
+import ud.ninia.prog3_proyecto.utils.posix.ShellCommand;
+
+public class Memory {
+
+    public static long getUsage() {
+        return Long.parseLong(
+                ShellCommand.executeCommand("/usr/local/proyectopbd/mem-usage.sh").replace("\n", "")
+        );
+    }
+
+    public static void main(String[] args) {
+        while(true){
+            System.out.println("Memory usage: " +
+                    "" + Memory.getUsage());
+            try {
+                Thread.sleep(1000);
+            } catch (InterruptedException ie) {
+                Thread.currentThread().interrupt();
+            }
+        }
+    }
+}
