+++ src/main/java/ud/ninia/prog3_proyecto/neo4j/omdb/MediaType.java
@@ -1,0 +1,5 @@
+package ud.ninia.prog3_proyecto.neo4j.omdb;
+
+public enum MediaType {
+    movie, series, episode, all;
+}
