+++ src/main/java/ud/ninia/prog3_proyecto/utils/network/HttpMethods.java
@@ -1,0 +1,8 @@
+package ud.ninia.prog3_proyecto.utils.network;
+
+/**
+ *  List of supported rest methods
+ */
+public enum HttpMethods {
+    GET, POST
+}
