+++ src/main/resources/web/planteamiento.png
