--- src/ud/main/influxdb/Log.java
@@ -1,92 +1,0 @@
-package ud.main.influxdb;
-
-
-import org.w3c.dom.Document;
-import ud.main.influxdb.monitor.CPU;
-import ud.main.influxdb.monitor.Memory;
-import ud.main.utils.DocumentReader;
-
-import java.util.ArrayList;
-
-/**
- *  Obtain data about the server contained in Points so it can be logged to InfluxDB
- */
-public class Log {
-
-    private static String SYSNAME = DocumentReader.getDoc(
-            "config/System.xml").getElementsByTagName("hostname").item(0).getTextContent();
-
-
-    /** Static method used for obtaining current CPU usage
-     *
-     * @return string containing current CPU usage percentage
-     */
-    public static Point getCPUUsagePoint(){
-
-        Point point = new Point();
-        point.setMeasurement("cpu_usage");
-        point.getTags().put("host", "\"" + SYSNAME + "\"");
-        point.getFields().put("value", CPU.getUsage());
-        point.setTime(System.currentTimeMillis());
-
-        return point;
-    }
-
-
-    /** Static method used for obtaining current CPU temperature
-     *
-     * @return string containing current CPU temperature percentage
-     */
-    public static Point getCPUTemperaturePoint(){
-
-        Point point = new Point();
-        point.setMeasurement("cpu_temperature");
-        point.getTags().put("host", "\"" + SYSNAME + "\"");
-        point.getFields().put("value", CPU.getTemperature());
-        point.setTime(System.currentTimeMillis());
-
-        return point;
-    }
-
-
-    /** Static method used for obtaining current memory usage
-     *
-     * @return string containing current memory usage percentage
-     */
-    public static Point getMemUsagePoint(){
-        Point point = new Point();
-        point.setMeasurement("mem_usage");
-        point.getTags().put("host", "\"" + SYSNAME + "\"");
-        point.getFields().put("value", Memory.getUsage());
-        point.setTime(System.currentTimeMillis());
-
-        return point;
-    }
-
-
-    /** Generates ArrayLists
-     *
-     * @param n of batch to be sent to InfluxDB
-     * @return ArrayList containing the generated points
-     */
-    public static ArrayList<Point> generatePoints(int n) {
-        ArrayList<Point> points = new ArrayList<>();
-        for (int i=0; i<n; i++){
-            points.add(getCPUUsagePoint());
-            points.add(getCPUTemperaturePoint());
-            points.add(getMemUsagePoint());
-        }
-        return points;
-    }
-
-
-    public static void main(String[] args) {
-        String db = "server_stats";
-        InfluxDB.createDataBase(db);
-        ArrayList<Point> points = generatePoints(10);
-        for (Point point: points) {
-            System.out.println(point);
-        }
-        InfluxDB.writePoints(db, points);
-    }
-}
