--- src/ud/main/influxdb/monitor/Memory.java
@@ -1,24 +1,0 @@
-package ud.main.influxdb.monitor;
-
-import ud.main.utils.posix.ShellCommand;
-
-public class Memory {
-
-    public static long getUsage() {
-        return Long.parseLong(
-                ShellCommand.executeCommand("/usr/local/proyectopbd/mem-usage.sh").replace("\n", "")
-        );
-    }
-
-    public static void main(String[] args) {
-        while(true){
-            System.out.println("Memory usage: " +
-                    "" + Memory.getUsage());
-            try {
-                Thread.sleep(1000);
-            } catch (InterruptedException ie) {
-                Thread.currentThread().interrupt();
-            }
-        }
-    }
-}
