--- src/ud/main/neo4j/omdb/MediaType.java
@@ -1,5 +1,0 @@
-package ud.main.neo4j.omdb;
-
-public enum MediaType {
-    movie, series, episode, all;
-}
