--- src/ud/main/neo4j/omdb/Omdb.java
@@ -1,70 +1,0 @@
-package ud.main.neo4j.omdb;
-
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.HashMap;
-import java.util.Scanner;
-
-/**
- * Java Implementation of the OMDB api
- * <p>
- * http://omdbapi.com/
- */
-public class Omdb {
-
-    private static String url = "";
-
-    /**
-     * Searches a title in IMDB, can differentiate between media types.
-     *
-     * @param title - Title to search for
-     * @param type  - Type of media to search for
-     * @return a HashMap where the key is the IMDBid
-     * and the value is another HashMap with basic information about the title
-     */
-    public static HashMap search(String title, MediaType type) {
-        try {
-
-            HashMap search_results = new HashMap<String, HashMap>();
-
-            if (!type.equals(MediaType.all))
-                url = "http://www.omdbapi.com/?s=" + title.replace(" ", "%20") + "&type=" + type.name();
-            else
-                url = "http://www.omdbapi.com/?s=" + title.replace(" ", "%20");
-
-            URL query = new URL(url);
-
-            Scanner s = new Scanner(query.openStream());
-            JSONObject response = new JSONObject(s.nextLine());
-
-            JSONArray results = response.getJSONArray("Search");
-
-            for (Object result : results) {
-                if (!result.equals("{}")) {
-
-                    JSONObject entry = new JSONObject(result.toString());
-
-                    HashMap entry_info = new HashMap<String, String>();
-                    entry_info.put("Year", entry.get("Year"));
-                    entry_info.put("Title", entry.get("Title"));
-                    entry_info.put("Type", entry.get("Type"));
-
-                    search_results.put(entry.get("imdbID"), entry_info);
-                }
-            }
-
-            return search_results;
-
-        } catch (MalformedURLException e) {
-            System.err.println("Malformed URL: " + url);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return null;
-    }
-}
