--- src/ud/main/utils/network/HttpMethods.java
@@ -1,8 +1,0 @@
-package ud.main.utils.network;
-
-/**
- *  List of supported rest methods
- */
-public enum HttpMethods {
-    GET, POST
-}
