--- src/ud/main/utils/posix/ShellCommand.java
@@ -1,29 +1,0 @@
-package ud.main.utils.posix;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-
-public class ShellCommand {
-    public static String executeCommand(String command) {
-
-        StringBuffer output = new StringBuffer();
-
-        Process p;
-        try {
-            p = Runtime.getRuntime().exec(command);
-            p.waitFor();
-            BufferedReader reader = new BufferedReader(new InputStreamReader(p.getInputStream()));
-
-            @SuppressWarnings("UnusedAssignment")
-            String line = "";
-            while ((line = reader.readLine())!= null) {
-                output.append(line + "\n");
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return output.toString();
-    }
-
-}
